backend/pyproject.toml
[project]
name = "be-ecofilia"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
package-mode = false
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "openai (>=1.97.1,<2.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-simplejwt[crypto] (>=5.4.0,<6.0.0)",
    "django-otp (>=1.5.4,<2.0.0)",
    "pyotp (>=2.9.0,<3.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "celery[sqs] (>=5.5.3,<6.0.0)",
    "boto3 (>=1.40.26,<2.0.0)",
    "django-storages (>=1.14.6,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
